+++ changelog.lua
@@ -22,6 +22,7 @@ local NO_CHANGELOG_VERSIONS = {
     ["0.5.8"] = true,
 	["0.5.10"] = true,
 	["0.5.11"] = true,
+	["0.5.18"] = true,
 }
 
 -- Changelog content (update this with each release)